#1188 Move FortranReader to a separate file
